Punto_venta/Backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
mysqlclient>=2.2
python-dotenv>=1.0
Pillow
pytest
pytest-django
pytest-cov
model-bakery
selenium>=4.0.0
webdriver-manager>=4.0.0
XlsxWriter>=3.2.0
reportlab>=4.0.9